1) Почему не могу воспльзоваться методами на имени list,хотя на имени созданного объекта могу. 2) В методах через цикл for бегаем по объектам списка militaryMans, он подсвечивается красным. не понятен этот момент.
